fix: 투표 페이지 버블차트 제거
